cookbook/llama-index/requirements.txt
llama-index-core>=0.10.0
llama-index-llms-openai-like>=0.1.0
llama-index-embeddings-huggingface>=0.1.0
python-dotenv>=1.0.0
